% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_D.R
\name{cal_D}
\alias{cal_D}
\title{calculate distance matrix}
\usage{
cal_D(data, cl)
}
\arguments{
\item{data}{transcripts}

\item{cl}{gene level cluster given by modal cluster}
}
\value{
distance matrix
}
\description{
calculate distance matrix
}
