% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calD.R
\name{calD}
\alias{calD}
\title{calculate distance matrix}
\usage{
calD(data, bp)
}
\arguments{
\item{data}{transcripts}

\item{bp}{bioc parallel parameter}
}
\value{
distance matrix
}
\description{
calculate distance matrix
}
\examples{
data(sim_dat)
dat <- extractInfo(sim_dat)
data_counts <- dat$count_matrix
bp <- BiocParallel::MulticoreParam(4)
D_c <- calD(data_counts,bp)
}
