% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detK.R
\name{clusHelper}
\alias{clusHelper}
\title{function to get intra and inter distance for clusters}
\usage{
clusHelper(D, i)
}
\arguments{
\item{D}{distance matrix}

\item{i}{number of clusters}
}
\value{
vector of intra and inter distance
}
\description{
function to get intra and inter distance for clusters
}
