% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detK.R
\name{detK}
\alias{detK}
\title{determine the number of clusters}
\usage{
detK(D, epi = 1)
}
\arguments{
\item{D}{distance matrix}

\item{epi}{threshold for cutting off}
}
\value{
number of clusters
}
\description{
determine the number of clusters
}
\examples{
data(sim_dat)
dat = extractInfo(sim_dat)
data_counts = dat$count_matrix
bp <- BiocParallel::MulticoreParam(4)
D_c = calD(data_counts,bp)
detK(D_c)
}
