% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBS.R
\name{EBS}
\alias{EBS}
\title{accelerated empirical bayesian}
\usage{
EBS(data, conditions, gclus, sf, iter = 10, hyper, PP, stp1, stp2)
}
\arguments{
\item{data}{single cell expression matrix, row as genes column as cells}

\item{conditions}{partition of cells}

\item{gclus}{partition of genes}

\item{sf}{size factors}

\item{iter}{maximum iteration step of EM}

\item{hyper}{hyper parameters for beta distributions}

\item{PP}{pattern of partitions}

\item{stp1}{step size of hyperparameter alpha (shared by all units) in one step EM}

\item{stp2}{step size of hyperparameter beta (unit specific) in one step EM}
}
\value{
posterior probability of mean expression pattern
}
\description{
accelerated empirical bayesian
}
