% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInfo.R
\name{extractInfo}
\alias{extractInfo}
\title{extract count matrix from SingleCellExperiment object}
\usage{
extractInfo(data)
}
\arguments{
\item{data}{SingleCellExperiment object}
}
\value{
list of count matrix and condition vector
}
\description{
extract count matrix from SingleCellExperiment object
}
