% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gCl.R
\name{gCl}
\alias{gCl}
\title{gene_level cluster}
\usage{
gCl(data, bp)
}
\arguments{
\item{data}{transcripts}

\item{bp}{bioc parallel parameter}
}
\value{
return a matrix whose row represent gene specific cluster
}
\description{
gene_level cluster
}
