% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genRClus.R
\name{genRClus}
\alias{genRClus}
\title{generate random clusterings}
\usage{
genRClus(D, a, K)
}
\arguments{
\item{D}{distance matrix of cells}

\item{a}{paramter for weights}

\item{K}{number of subtypes}
}
\value{
random generated clustering of cells
}
\description{
generate random clusterings
}
