% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSizeofDD.R
\name{getSizeofDD}
\alias{getSizeofDD}
\title{number of DD genes under FDR control}
\usage{
getSizeofDD(pDD, FDR = 0.01)
}
\arguments{
\item{pDD}{estimated probability of being DD}

\item{FDR}{fdr to be controlled}
}
\value{
number of positive genes
}
\description{
number of DD genes under FDR control
}
\examples{
p_dd = c(0.1,0.99,1,0.05,0.05)
getSizeofDD(p_dd)
}
