% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdd.R
\name{getZ1Z2}
\alias{getZ1Z2}
\title{function to get counts of cluster sizes at two conditions}
\usage{
getZ1Z2(ccl, cd)
}
\arguments{
\item{ccl}{clustering label}

\item{cd}{condition label}
}
\value{
return list of counts
}
\description{
function to get counts of cluster sizes at two conditions
}
