% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRef.R
\name{gRef}
\alias{gRef}
\title{generate reference matrix}
\usage{
gRef(Posp)
}
\arguments{
\item{Posp}{possible partition of data}
}
\value{
return a matrix indicate the refinement relation between different partitions.
}
\description{
generate reference matrix
}
