% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isRef.R
\name{isRef}
\alias{isRef}
\title{check refinement relation between two clusters}
\usage{
isRef(x, y)
}
\arguments{
\item{x}{a cluster}

\item{y}{a cluster}
}
\value{
whether x refines y
}
\description{
check refinement relation between two clusters
}
