% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDD.R
\name{lpt1t2}
\alias{lpt1t2}
\title{log likelihood of z1,z2 given t1,t2}
\usage{
lpt1t2(z1, z2, pp, alpha1, alpha2)
}
\arguments{
\item{z1}{counts of each group in condition 1}

\item{z2}{counts of each group in condition 2}

\item{pp}{a partition}

\item{alpha1}{parameter of double dirichlet prior}

\item{alpha2}{parameter of double dirichlet prior}
}
\value{
log likelihood of z1,z2 given t1,t2
}
\description{
log likelihood of z1,z2 given t1,t2
}
