% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDD.R
\name{lpzgt}
\alias{lpzgt}
\title{log likelihood of aggregated multinomial counts z given aggregated proportions t}
\usage{
lpzgt(z, pp, alpha)
}
\arguments{
\item{z}{counts of each group in one condition}

\item{pp}{a partition}

\item{alpha}{parameter of double dirichlet prior}
}
\value{
log likelihood of aggregated multinomial counts z given aggregated proportions t
}
\description{
log likelihood of aggregated multinomial counts z given aggregated proportions t
}
