% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDD_random.R
\name{PDD_random}
\alias{PDD_random}
\title{calculate PDD when add random noise in distance matrix}
\usage{
PDD_random(data, cd, K, D, hp, Posp, iter, lambda)
}
\arguments{
\item{data}{normalized preprocessed transcripts}

\item{K}{number of subgroups}

\item{D}{distance matrix of cells}

\item{hp}{hyper parameters for EBSeq}

\item{Posp}{parition patterns}

\item{iter}{max number of iterations for EM}

\item{lambda}{parameter for random noise}

\item{ncores}{number cores for parallel computing}
}
\value{
posterior probabilities under random distance matrix
}
\description{
calculate PDD when add random noise in distance matrix
}
