% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdd.R
\name{pddAggregate}
\alias{pddAggregate}
\title{function to aggregate intermediate results and get prob of DD}
\usage{
pddAggregate(z1, z2, Posp, DE, K, REF)
}
\arguments{
\item{z1}{counts of cluster sizes in condition 1}

\item{z2}{counts of cluster sizes in condition 2}

\item{Posp}{partition of cells}

\item{DE}{posterior probabilities of DE patterns}

\item{K}{number of clusters}

\item{REF}{reference matrix indicating relation of nested partitions}
}
\value{
return vector of prob of DD
}
\description{
function to aggregate intermediate results and get prob of DD
}
