% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scToMatrix.R
\name{scToMatrix}
\alias{scToMatrix}
\title{extract count matrix from SingleCellExperiment object}
\usage{
scToMatrix(data)
}
\arguments{
\item{data}{SingleCellExperiment object}
}
\value{
count matrix
}
\description{
extract count matrix from SingleCellExperiment object
}
