\name{scDDboost-package}
\alias{scDDboost-package}
\alias{scDDboost}
\docType{package}
\title{
\packageTitle{scDDboost}
}
\description{
\packageDescription{scDDboost}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scDDboost}
\packageIndices{scDDboost}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{scDDboost}

Maintainer: \packageMaintainer{scDDboost}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
